Match Inheritance file name with exported class name.
